Update copyright and URL references in multiple files
Changes have been made to the copyright texts and URL references found in several files. The updated copyright texts now include a URL link, while the source references in the quote sections have been shortened and modified to link to a more specific page. Furthermore, adjustments were made to the aliases sections in the front matter of various posts to correct potential issues with URL pathing.