Update project module name and homepage content
The project module name has been updated from 'marskingx.github.io.iml' to a new name. In addition, the homepage content in 'call-to-action-home.md' has been significantly rewritten. This includes changing the description to a more comprehensive text in Chinese and updating the button label and link to lead to the blog page.